moved code to datapath_cc::emitMux()
